Include station numbers in WCIF when enabled on scorecards
Currently there is no proper support for fixed stations, however we
have an option to print station number on scorecards by generating
the numbers on the fly. This change adds station numbers to competitor
assignments in WCIF, instead of generating them on the fly, so they
can be consumed by other software. The user is still responsible for
making sure the groups are small enough, such that everybody has a
station.