Add a star to the upper-right of some scorecards (#61)
This adds an option to the ActivityConfig WCIF extension; when this is
set, the specified competitors will have a star in the upper right. This
may be used in major competitions to indicate which competitors should
solve on camera for the live stream.

This doesn't provide any logic for marking competitors as featured. For
CubingUSA Nationals we'll be using another program for generating
groups, and Groupifier scorecards. If other competitions would find this
useful, a UI could be built to mark certain competitors as featured, but
I'm not sure how often this would be useful.